rename geometry._helpers into geometry._check, move deprecated into geometry._docs, move pyriemann._base into utils
