+++ pyriemann/geometry/distance.py
@@ -8,8 +8,9 @@ from array_api_compat import (
 from sklearn.metrics import euclidean_distances
 
 from ._backend import diag_indices, torch, tril_indices
-from ._helpers import check_function, check_matrix_pair, is_real_type
+from ._check import check_function, check_matrix_pair
 from .base import _eigvalsh, invsqrtm, logm, powm, sqrtm
+from .test import is_real_type
 
 
 ###############################################################################
