+++ pyriemann/geometry/test.py
@@ -1,7 +1,8 @@
-from array_api_compat import array_namespace as get_namespace
+from array_api_compat import (
+    array_namespace as get_namespace,
+    is_torch_namespace,
+)
 import numpy as np
-
-from ._helpers import is_real_type, is_square  # noqa: F401
 
 
 def _allclose(A, B):
@@ -20,6 +21,22 @@ def _get_eigenvals(X):
     xp = get_namespace(X)
     n = X.shape[-1]
     return xp.real(xp.linalg.eigvals(X.reshape((-1, n, n))))
+
+
+def is_square(X):
+    """Check if matrices are square.
+
+    Parameters
+    ----------
+    X : ndarray, shape (..., n, n)
+        The set of square matrices, at least 2D ndarray.
+
+    Returns
+    -------
+    ret : bool
+        True if matrices are square.
+    """
+    return X.ndim >= 2 and X.shape[-2] == X.shape[-1]
 
 
 def is_sym(X):
@@ -103,6 +120,29 @@ def is_real(X):
     xp = get_namespace(X)
     X_imag = xp.imag(X)
     return _allclose(X_imag, xp.zeros_like(X_imag))
+
+
+def is_real_type(X):
+    """Check if matrices are real type.
+
+    Parameters
+    ----------
+    X : ndarray, shape (..., n, m)
+        The set of matrices.
+
+    Returns
+    -------
+    ret : bool
+        True if matrices are real type.
+
+    Notes
+    -----
+    .. versionadded:: 0.6
+    """
+    xp = get_namespace(X)
+    if is_torch_namespace(xp):
+        return not X.dtype.is_complex
+    return np.isrealobj(X)
 
 
 def is_hermitian(X):
