+++ pyriemann/utils/_base.py
