--- pyriemann/utils/docs.py
@@ -1,5 +1,0 @@
-"""Backward-compatibility re-export.
-
-The ``deprecated`` decorator now lives in :mod:`pyriemann.geometry._helpers`.
-"""
-from ..geometry._helpers import deprecated  # noqa: F401
